Moved xml report formatters & updated test result fitlers
