+++ course/utils/junit_xml.py
@@ -1,5 +1,6 @@
 import logging
-from course.utils.utils import format_message, format_test_name
+import re
+
 
 from bs4 import BeautifulSoup
 
@@ -29,3 +30,18 @@ def parse_junit_xml(xml):
         logger.error(e)
 
     return results
+
+
+def format_message(message):
+    starting_arrow_index = message.find("==>")
+    return message if starting_arrow_index == -1 else message[:starting_arrow_index]
+
+
+def convert_camel_case_to_title_case(text):
+    title_case_str = re.sub('(.)([A-Z][a-z]+)', r'\1 \2', text)
+    return re.sub('([a-z0-9])([A-Z0-9])', r'\1 \2', title_case_str)
+
+
+def format_test_name(name):
+    title_case_str = convert_camel_case_to_title_case(name)
+    return title_case_str.replace('()', '').capitalize()
