+++ course/utils/utils.py
@@ -1,6 +1,3 @@
-import re
-
-
 def get_user_question_junction(user, question):
     from course.models.models import UserQuestionJunction
 
@@ -178,18 +175,3 @@ def create_java_question(pk=None, title=None, text=None, max_submission_allowed=
             event=event,
         )
         question.save()
-
-
-def format_message(message):
-    starting_arrow_index = message.find("==>")
-    return message if starting_arrow_index == -1 else message[:starting_arrow_index]
-
-
-def convert_camel_case_to_title_case(text):
-    title_case_str = re.sub('(.)([A-Z][a-z]+)', r'\1 \2', text)
-    return re.sub('([a-z0-9])([A-Z0-9])', r'\1 \2', title_case_str)
-
-
-def format_test_name(name):
-    title_case_str = convert_camel_case_to_title_case(name)
-    return title_case_str.replace('()', '').capitalize()
